Make linger_ms user-configurable
For some reason when running using AsyncClient against the app directly
it is pausing for linger_ms before .send() is returning, even though
this is not meant to wait. This is pretty strange so I want a way to
modify it in the tests, and potentially test on production to see if it
is causing issues there.